Extracted create_dataset_from_snapshot from dataset repository interface, promoted to free shared algorithm
